[Game][Login] Added title picture [Game] Added SQL auto-update [Game] Updated to the current version of aaemu.develop 1.2 [Game] Entrance to the world

+++ AAEmu.Game/Core/Managers/AccountManager.cs
@@ -28,7 +28,7 @@ namespace AAEmu.Game.Core.Managers
             _accounts.TryRemove(id, out _);
         }
 
-        public bool Contains(uint id)
+        public bool Contains(ulong id)
         {
             return _accounts.ContainsKey(id);
         }
